organise methods inside utils.py into separate modules
