+++ setup.cfg
@@ -18,6 +18,7 @@ packages =
     ethereum/frontier
     ethereum/frontier/vm
     ethereum/frontier/vm/instructions
+    ethereum/utils
 
 package_dir =
     =src
