+++ src/ethereum/genesis.py
@@ -20,7 +20,7 @@ from typing import Dict, cast
 from ethereum.base_types import U256, Bytes, Bytes8, Uint, slotted_freezable
 from ethereum.frontier.eth_types import Address
 
-from .utils import (
+from .utils.hexadecimal import (
     hex_to_address,
     hex_to_bytes,
     hex_to_bytes8,
