+++ src/ethereum/utils/__init__.py
@@ -1,0 +1,13 @@
+"""
+Utility Functions
+^^^^^^^^^^^^^^^^^
+
+..contents:: Table of Contents
+    :backlinks: none
+    :local:
+
+Introduction
+------------
+
+Utility functions used in this application.
+"""
