+++ tests/frontier/blockchain_st_test_helpers.py
@@ -13,7 +13,7 @@ from ethereum.frontier.eth_types import (
     Transaction,
 )
 from ethereum.frontier.spec import BlockChain, state_transition
-from ethereum.utils import (
+from ethereum.utils.hexadecimal import (
     hex_to_address,
     hex_to_bytes,
     hex_to_bytes8,
@@ -22,8 +22,8 @@ from ethereum.utils import (
     hex_to_root,
     hex_to_u256,
     hex_to_uint,
-    rlp_hash,
 )
+from ethereum.utils.rlp import rlp_hash
 
 TEST_DIR = (
     "tests/fixtures/LegacyTests/Constantinople/BlockchainTests/"
